chore: deletes runtime and moves to react pkg
