+++ packages/react/runtime/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "../dist/runtime.js",
+  "module": "../dist/runtime.js",
+  "browser": "./dist/browser/index.js",
+  "types": "../dist/runtime.d.ts"
+}
