+++ packages/react/src/index.tsx
@@ -1,4 +1,6 @@
-export { CS, CC, ax } from '@compiled/runtime';
+// TODO: REMOVE THIS EXPORT
+export { CS, CC, ax } from './runtime';
+// END TODO
 export { styled } from './styled';
 export { ClassNames } from './class-names';
 import { CssFunction } from './types';
