+++ packages/react/src/runtime/is-node.tsx
