+++ packages/react/tsconfig.browser.json
