--- packages/runtime/README.md
@@ -1,3 +1,0 @@
-# @compiled/runtime
-
-The _actual_ runtime for compiled.
