--- packages/runtime/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@compiled/runtime",
-  "version": "0.4.7",
-  "description": "The CSS in JS authoring experience you love without the runtime cost",
-  "author": "Michael Dougall",
-  "license": "Apache-2.0",
-  "homepage": "https://compiledcssinjs.com",
-  "bugs": "https://github.com/atlassian-labs/compiled-css-in-js/issues/new?assignees=&labels=bug&template=bug_report.md",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/atlassian-labs/compiled-css-in-js.git",
-    "directory": "packages/runtime"
-  },
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "browser": "./dist/browser/index.js",
-  "types": "./dist/index.d.ts",
-  "sideEffects": false,
-  "files": [
-    "dist",
-    "src",
-    "README.md"
-  ],
-  "peerDependencies": {
-    "react": "^16.12.0"
-  }
-}
