--- packages/runtime/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "../tsconfig.options.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "baseUrl": "src",
-    "paths": {
-      "@compiled/runtime": ["index.tsx"]
-    },
-    "plugins": [
-      {
-        "transform": "ts-transform-define",
-        "replace": {
-          "isNodeEnvironment()": "process.env.IS_NODE_EXPRESSION"
-        }
-      }
-    ]
-  },
-  "references": []
-}
