+++ packages/tsconfig.json
@@ -5,7 +5,6 @@
     { "path": "css/tsconfig.json" },
     { "path": "react/tsconfig.json" },
     { "path": "jest/tsconfig.json" },
-    { "path": "runtime/tsconfig.json" },
     { "path": "utils/tsconfig.json" }
   ]
 }
